+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dotenv-run/core
 
+## 1.2.1
+
+### Patch Changes
+
+- refactor(esbuild): usde esbuild define option
+
 ## 1.2.0
 
 ### Minor Changes